feat: 기존 분리된 페이지 라우팅 제거 후 통합, SSR페이지 내 CSR컴포넌트 구조로 변경 #217
